requirements.txt
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
dnacentersdk==2.4.2
et-xmlfile==1.1.0
fastjsonschema==2.15.1
future==0.18.2
idna==3.3
numpy==1.21.4
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pip-review==1.1.0
pydantic==1.8.2
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
requests-toolbelt==0.9.1
six==1.16.0
typing_extensions==4.0.1
urllib3==1.26.7
xlrd==2.0.1
